# lib/api-zod/src/generated/types/createDebtRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wasi - Islamic Digital Will & Heritage Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDebtRequestType } from "./createDebtRequestType";

export interface CreateDebtRequest {
  type: CreateDebtRequestType;
  description: string;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  creditorName?: string | null;
  /** @nullable */
  creditorContact?: string | null;
}