# lib/api-zod/src/generated/types/digitalAsset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wasi - Islamic Digital Will & Heritage Management API
 * OpenAPI spec version: 0.1.0
 */
import type { DigitalAssetAction } from "./digitalAssetAction";

export interface DigitalAsset {
  id: string;
  userId: string;
  platform: string;
  /** @nullable */
  accountIdentifier?: string | null;
  instructions: string;
  action: DigitalAssetAction;
  /** @nullable */
  encryptedCredentials?: string | null;
  createdAt: string;
}