# lib/api-zod/src/generated/types/guardian.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wasi - Islamic Digital Will & Heritage Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Guardian {
  id: string;
  userId: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  relationship: string;
  isConfirmed: boolean;
  createdAt: string;
}