# lib/api-zod/src/generated/types/updateGuardianRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wasi - Islamic Digital Will & Heritage Management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateGuardianRequest {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  relationship?: string | null;
}