# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wasi - Islamic Digital Will & Heritage Management API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  lastCheckinAt?: string | null;
  checkinIntervalDays: number;
  createdAt: string;
}