# lib/api-zod/src/generated/types/will.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wasi - Islamic Digital Will & Heritage Management API
 * OpenAPI spec version: 0.1.0
 */
import type { WillStatus } from "./willStatus";

export interface Will {
  id: string;
  userId: string;
  title: string;
  content: string;
  isEncrypted: boolean;
  status: WillStatus;
  createdAt: string;
  updatedAt: string;
}